cleaned up transformer
