+++ mach/src/platform/transformation/mod.rs
@@ -1,3 +1,7 @@
 mod resolve_and_transform;
+mod resolve_and_transform_worker;
+mod run_graph_update;
+mod run_resolvers;
+mod run_transformers;
 
 pub use self::resolve_and_transform::*;
